+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - ADDED: LAN/Wi-Fi global IPv6 addresses are now excluded when configuring firewall rules.
 - ADDED: `IPsec` VPN subnet exclusion – skips subnet rule when IPsec is disabled.
 - FIXED: Don't resolve log ips when dnsmasq is disabled.
+- FIXED: Correctly resolve online a device object by an ip6 in the logs.
 
 ## [0.52.2] - 2025-07-13